backend/requirements.txt
# API
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.8

# DB & queue
sqlmodel>=0.0.22
psycopg[binary]>=3.2
alembic>=1.13
redis>=5.0
rq>=1.16

# Core agent
pandas>=2.2
pyyaml>=6.0
python-dotenv>=1.0
requests>=2.32
rapidfuzz>=3.9

# Exports
openpyxl>=3.1
weasyprint>=62.3

# Auth (optionnel)
pyjwt>=2.9
passlib[bcrypt]>=1.7